requirements.txt
# 주식 데이터 수집
yfinance>=0.2.36
yahooquery>=2.3.7

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 기술적 지표는 직접 구현 (pandas, numpy 사용)
# ta 라이브러리는 선택사항 (주석 처리)
# ta>=0.11.0

# API 클라이언트
requests>=2.31.0
fredapi>=0.5.0

# 텔레그램
python-telegram-bot>=20.0

# 유틸리티
python-dotenv>=1.0.0
pytz>=2024.1

# 로깅 및 에러 핸들링
tenacity>=8.2.0
